Stop using C++ streams
This is mainly an attempt to reduce binary size. Doing this reduces internal sizes significantly: the final binary size decreases only slightly, from 1976960 bytes to 1951488 bytes, an almost 25 kB reduction.

+++ main/kernel/Kernel.hpp
@@ -104,11 +104,11 @@ public:
     }
 
     void prepareUpdate(const std::string& url) {
-        std::ofstream fUpdate = fs.openWrite(UPDATE_FILE);
         JsonDocument doc;
         doc["url"] = url;
-        serializeJson(doc, fUpdate);
-        fUpdate.close();
+        std::string content;
+        serializeJson(doc, content);
+        fs.writeAll(UPDATE_FILE, content);
     }
 
     void performFactoryReset(bool completeReset) {
@@ -220,10 +220,12 @@ private:
         PowerManagementLockGuard sleepLock(preventLightSleep);
 #endif
 
-        std::ifstream fUpdate = fs.openRead(UPDATE_FILE);
+        auto contents = fs.readAll(UPDATE_FILE);
+        if (!contents.has_value()) {
+            return "Failed to read update file";
+        }
         JsonDocument doc;
-        auto error = deserializeJson(doc, fUpdate);
-        fUpdate.close();
+        auto error = deserializeJson(doc, contents.value());
         unlink(UPDATE_FILE);
 
         if (error) {
